+++ src/components/bookmark.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Bookmark = ({status,...rest}) => {
+
+    return
+
+}
+
+export default Bookmark
