+++ src/components/quality.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Quality = ({color,name,_id}) => {
+
+return
+
+}
+
+export default Quality
