+++ src/components/searchStatus.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const SearchStatus = ({length}) => {
+
+
+    return (<> <span className={getBadgeClasses()}>{renderPhrase(users.length)}</span></>
+    )
+}
+
+export default SearchStatus
