+++ src/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import reactDOM from 'react-dom';
 import "bootstrap/dist/css/bootstrap.css"
-import Users from "./components/users";
 import reportWebVitals from "./reportWebVitals";
-
+import App from "./app";
 // const App = () => {
 //     return <Users/>
 // }
@@ -13,7 +12,7 @@ import reportWebVitals from "./reportWebVitals";
 // )
 reactDOM.render(
   <React.StrictMode>
-    <Users />
+    <App />
   </React.StrictMode>,
   document.getElementById('root')
 );
